Extract tests for isMap to own file
